♻️ Refactor and modularize NavBar
