+++ src/components/Navbar/index.ts
@@ -1,0 +1,3 @@
+export * from './Navbar';
+export * from './NavItems';
+export * from './SocialIcons';
